requirements.txt
appnope==0.1.0
argon2-cffi==20.1.0
astroid==2.4.2
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.6.0
bleach==3.1.5
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docopt==0.6.2
entrypoints==0.3
idna==2.10
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
JPype1==1.0.2
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
kiwisolver==1.2.0
konlpy==0.5.2
lazy-object-proxy==1.4.3
lxml==4.5.2
MarkupSafe==1.1.1
matplotlib==3.3.0
mccabe==0.6.1
mistune==0.8.4
mpmath==1.1.0
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.1.1
numpy==1.19.1
oauthlib==3.1.0
packaging==20.4
pandas==1.1.0
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
praat-parselmouth==0.3.3
prometheus-client==0.8.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pyannote.core==4.1
pyannote.database==4.0.1
pyannote.metrics==3.0.1
pycparser==2.20
Pygments==2.6.1
pylangacq==0.11.0
pylint==2.5.3
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
requests==2.24.0
requests-oauthlib==1.3.0
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.10.1
Send2Trash==1.5.0
shellingham==1.3.2
simplejson==3.17.2
six==1.15.0
sortedcontainers==2.2.2
sympy==1.6.1
tabulate==0.8.7
terminado==0.8.3
testpath==0.4.4
TextGrid==1.5
threadpoolctl==2.1.0
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
tweepy==3.9.0
typer==0.3.1
typing-extensions==3.7.4.2
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.12.1
